+++ client/row_actions/undoRedo.js
@@ -23,6 +23,8 @@ Template.Constellation_undoRedo.events({
   }
 });
 
+EditableJSON = Package['babrahams:editable-json'].EditableJSON;
+
 if (!!EditableJSON) {
   EditableJSON.afterUpdate(function(collection, action, documentBefore, result) {
     /*console.log("doc before:",documentBefore);